Move sync schema fn to helper
